[FIX] 품앗이 화면 세그먼트 제거
