--- DuruDuru/Exchange(품앗이)/VIews/MainExchangeView.swift
@@ -1,104 +1,0 @@
-//
-//  ExchangeView.swift
-//  DuruDuru
-//
-//  Created by 임효진 on 1/23/25.
-//
-
-import UIKit
-
-class MainExchangeView: UIView {
-    
-    // MARK: - Init
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addComponents()
-        constraints()
-        
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - Components
-    
-    /// 주소 바꾸기 버튼
-    let locationButton = UIButton().then {
-        $0.setTitle("내 동네 설정", for: .normal)
-        $0.setTitleColor(.black, for: .normal)
-        $0.titleLabel?.font = UIFont.boldSystemFont(ofSize: 15)
-    }
-    
-    let downImage = UIImageView().then {
-        $0.image = UIImage(systemName: "chevron.down")?.withRenderingMode(.alwaysTemplate) 
-            $0.tintColor = .black
-        $0.contentMode = .scaleAspectFit
-        $0.frame = CGRect(x: 0, y: 0, width: 16, height: 16)
-    }
-    
-    /// 세그먼트
-//    let segmentedControl = UISegmentedControl(items: ["품앗이", "함께 먹자"]).then {
-//        $0.setBackgroundImage(UIImage(), for: .normal, barMetrics: .default)
-//        $0.setBackgroundImage(UIImage(), for: .selected, barMetrics: .default)
-//        $0.setBackgroundImage(UIImage(), for: .highlighted, barMetrics: .default)
-//        $0.setDividerImage(UIImage(), forLeftSegmentState: .selected, rightSegmentState: .normal, barMetrics: .default)
-//        $0.selectedSegmentIndex = 0
-//        $0.setTitleTextAttributes(
-//            [
-//                NSAttributedString.Key.foregroundColor: UIColor.black,
-//                .font: UIFont.systemFont(ofSize: 14, weight: .light),
-//            ],
-//            for: .normal
-//        )
-//        $0.setTitleTextAttributes(
-//            [
-//                NSAttributedString.Key.foregroundColor: UIColor.black,
-//                .font: UIFont.systemFont(ofSize: 14, weight: .bold),
-//            ],
-//            for: .selected
-//        )
-//    }
-    
-    /// 세그먼트 밑줄
-//    let underline = UIView().then {
-//        $0.backgroundColor = .white
-//    }
-    
-    let containerView = UIView()
-    
-    // MARK: - Constaints & Add Function
-    
-    /// 컴포넌트 생성
-    private func addComponents() {
-//        addSubview(segmentedControl)
-//        addSubview(underline)
-        addSubview(containerView)
-    }
-    
-    /// 오토레이아웃 설정
-    private func constraints() {
-        
-//        segmentedControl.snp.makeConstraints {
-//            $0.top.equalToSuperview().offset(88)
-//            $0.centerX.equalToSuperview()
-//            $0.width.equalToSuperview()
-//            $0.height.equalTo(48)
-//        }
-//        
-//        underline.snp.makeConstraints {
-//            $0.height.equalTo(1)
-//            $0.width.equalTo(segmentedControl.snp.width).dividedBy(segmentedControl.numberOfSegments)
-//            $0.top.equalTo(segmentedControl.snp.bottom).offset(3)
-//            $0.left.equalTo(segmentedControl.snp.left)
-//        }
-        
-        containerView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(100)
-            $0.leading.trailing.bottom.equalToSuperview()
-        }
-        
-    }
-
-}
